feat: move check_nodes_number from node_pool_util to jobset_ttr_restart dag
